Improve caching of Doom Emacs
Use another stage to install Doom Emacs on a basic Ubuntu image. Then
copy the files installed to the target base image. This way if the base
image changes, Doom Emacs doesn't need to be installed again. Which is
good because it takes forever and occasionally fails.

It's important to note that the Straight package manager used by Doom
makes heavy use of symlinks with absolute paths. (Why didn't they use
relative paths?) Copying those symlinks to a new location causes them to
break, unless the absolute paths are identical in the cache layer and
target image. So the cache layer was set up to match.